Move async methods to Utils
